+++ src/main/java/yesman/epicfight/api/client/event/instances/package-info.java
@@ -1,6 +1,6 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-package yesman.epicfight.api.hook.instances;
+package yesman.epicfight.api.client.event.instances;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;