+++ src/main/java/yesman/epicfight/api/hook/EpicFightHooks.java
@@ -1,5 +1,0 @@
-package yesman.epicfight.api.hook;
-
-public final class EpicFightHooks {
-	private EpicFightHooks() {}
-}