+++ src/main/java/yesman/epicfight/api/hook/subscriptions/HookSubscription.java
@@ -1,6 +1,0 @@
-package yesman.epicfight.api.hook.subscriptions;
-
-import yesman.epicfight.api.hook.EventInstance;
-
-public interface HookSubscription<T extends EventInstance> {
-}