+++ src/main/java/yesman/epicfight/skill/SkillCategories.java
@@ -19,13 +19,14 @@ public enum SkillCategories implements SkillCategory {
 	final boolean shouldSyncronize;
 	final boolean modifiable;
 	final int id;
-	ResourceLocation bookIcon;
+	final ResourceLocation bookIcon;
 	
 	SkillCategories(boolean shouldSave, boolean shouldSyncronizedAllPlayers, boolean modifiable) {
 		this.shouldSave = shouldSave;
 		this.shouldSyncronize = shouldSyncronizedAllPlayers;
 		this.modifiable = modifiable;
 		this.id = SkillCategory.ENUM_MANAGER.assign(this);
+		this.bookIcon = SkillCategory.DEFAULT_BOOK_ICON;
 	}
 	
 	SkillCategories(boolean shouldSave, boolean shouldSyncronizedAllPlayers, boolean modifiable, ResourceLocation bookIcon) {
@@ -58,6 +59,6 @@ public enum SkillCategories implements SkillCategory {
 	
 	@Override
 	public ResourceLocation bookIcon() {
-		return this.bookIcon == null ? SkillCategory.DEFAULT_BOOK_ICON : this.bookIcon;
+		return this.bookIcon;
 	}
 }