+++ RELEASE_NOTES.md
@@ -36,6 +36,7 @@
     - We will eventually replace all Forge/Neoforge events owned by Epic Fight into Events
 - Rename the experimental enum `EpicFightInputActions` to `EpicFightInputAction` to follow Java naming
   conventions. [#2194](https://github.com/Epic-Fight/epicfight/issues/2194)
+- Removed AirAttack and its related fields (SkillCategory, SkillSlot) to merge air slash and combo attacks as one skill
 
 ## [20.13.6] - 2025-11-12